refactored addForm component
