+++ data/db.json
@@ -21,19 +21,9 @@
       "id": 4
     },
     {
-      "name": "hodal data",
-      "description": "dasdadasdadadad",
+      "name": "testing 01",
+      "description": "Pellentesque in ipsum id orci porta dapibus. Quisq",
       "id": 5
-    },
-    {
-      "name": "Andela Task",
-      "description": "complete the react task",
-      "id": 6
-    },
-    {
-      "name": "coding",
-      "description": "fffffffffffffffffffffffff",
-      "id": 7
     }
   ]
 }
